LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpEncodingParameters-codec-opus-stereo.https-expected.txt


FAIL Sending and receiving a mono track with opus (stereo decoder upmix) assert_equals: Can open track in mono expected (number) 1 but got (undefined) undefined
FAIL Sending and receiving a stereo track with opus assert_equals: Can open track in stereo expected (number) 2 but got (undefined) undefined
